refactor: 내비게이션 바 UI 컴포넌트 FSD 구조에 맞게 경로 변경
