--- src/features/navigation/index.ts
@@ -1,3 +1,0 @@
-export { default as LoginButton } from './ui/LoginButton';
-export { default as NavigationTabs } from './ui/NavigationTabs';
-export { default as UserDropdown } from './ui/UserDropdown';
