+++ src/widgets/navigation-bar/NavigationBar.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import { Logo, NavigationMenu, DropdownMenu, DropdownMenuTrigger } from '@/shared';
-import { NavigationTabs, LoginButton, UserDropdown } from '@/features/navigation';
-import { useUserStore, UserIcon } from '@/entities/user';
+import { NavigationTabs, LoginButton, UserDropdown, UserIcon } from './ui';
+import { useUserStore } from '@/entities/user';
 import { useShallow } from 'zustand/shallow';
 
 export default function NavigationBar() {
@@ -14,7 +14,7 @@ export default function NavigationBar() {
   );
   // const isLoggedIn = true;
   return (
-    <div className="w-full bg-white sticky flex justify-center border-b border-[#E3E7EB]">
+    <div className="w-full bg-white flex justify-center border-b border-[#E3E7EB] sticky top-0">
       <div className="flex justify-between h-[75px] w-7xl items-center px-5">
         <NavigationMenu className="gap-6">
           <Logo />
