+++ src/widgets/navigation-bar/ui/LoginButton.tsx
