+++ src/widgets/navigation-bar/ui/NavigationTabs.tsx
