+++ src/widgets/navigation-bar/ui/UserDropdown.tsx
