+++ src/widgets/navigation-bar/ui/index.ts
@@ -1,0 +1,4 @@
+export { default as LoginButton } from './LoginButton';
+export { default as NavigationTabs } from './NavigationTabs';
+export { default as UserDropdown } from './UserDropdown';
+export { default as UserIcon } from './UserIcon';
